changed pathways and concentration from string to set of strings
Since the same course can be in multiple pathways and concentrations.

Other classes that directly accessed the concentration and pathway data has also been updated.

I will be adding in functions to help manage the data in the future so that updating the representation of any value will not affect anything else.