degree requirement checking algorithm prototype.
Degree checking algorithm prototype completed:

A course_template specifies required fields with the option to use wildcards (*), and an optional course_set to specify the pool of courses to draw from.

get_course_match inside catalog.py then finds all appropriate matching courses and returns it back as a set if no wildcards are used, or a dictionary if wildcards are used. The dictionary's keys are combinations that are generated from the wildcards.

Honestly this bitch was way overengineered for what it needs to be but I'm a sucker for generalized algorithms lmao