new parsing algorithm o(≧口≦)o
Completely rewrote the parser:

1) gets user command input and parse it into Command objects containing a command and arguments.

2) push each separate Command into a queue stored inside the user object. If queue is not done executing, we block any further command inputs

3) we execute the command one by one, unless a command requests user input. If that is the case, we add the command to a temporary storage area within the user object because we can't add it back to the stupid queue.

We make a note of this by setting the CMD_PAUSED flag inside the user, and if we see this, we will hand over the next user input to another temporary variable inside user.

We then execute the while loop except instead of getting from the queue, we use the last command and attempt to complete that command with the knowledge of the new user input. This cycle may continue for as long as necessary until the user input fulfills the command and everything keeps moving.

4) every time the while loop finishes a cycle where an element was gotten from the queue, make sure to use queue.task_done() to let the queue object know a task was completed, this is how the queue keeps track of the queue lock, since we block command input (we put it on hold) until queue gets unblocked.

We need to do this because otherwise it is possible for simultaneous queue solvers to run at the same time which may cause duplicated execution and a bunch of stuff that will mess us up. Queue execution at the moment must be single threaded, and I don't think multithreading it will be necessary since different users can be (and currently are being) multithreaded instead.